+++ examples/counter.exs
@@ -10,12 +10,12 @@ alias ExRatatui.Event
 
 defmodule Counter do
   def run do
-    ExRatatui.run(fn ->
-      loop(0)
+    ExRatatui.run(fn terminal ->
+      loop(terminal, 0)
     end)
   end
 
-  defp loop(count) do
+  defp loop(terminal, count) do
     {w, h} = ExRatatui.terminal_size()
 
     paragraph = %Paragraph{
@@ -24,20 +24,20 @@ defmodule Counter do
       alignment: :center
     }
 
-    ExRatatui.draw([{paragraph, %Rect{x: 0, y: 0, width: w, height: h}}])
+    ExRatatui.draw(terminal, [{paragraph, %Rect{x: 0, y: 0, width: w, height: h}}])
 
     case ExRatatui.poll_event(100) do
       %Event.Key{code: "q", kind: "press"} ->
         :ok
 
       %Event.Key{code: code, kind: "press"} when code in ["up", "k"] ->
-        loop(count + 1)
+        loop(terminal, count + 1)
 
       %Event.Key{code: code, kind: "press"} when code in ["down", "j"] ->
-        loop(count - 1)
+        loop(terminal, count - 1)
 
       _ ->
-        loop(count)
+        loop(terminal, count)
     end
   end
 end